Make the FDW work in PostgreSQL 9.2
That is a simple solution to make the mongo_fdw work on PostgreSQL 9.2.
There is no performance improvements that the 9.2 version provided, like
better estimate plans. Basically we created the MongoGetForeignRelSize
function based on old MongoPlanForeignScan (encapsulating the calls on
MongoGeneratePlanState function), and basic (dummy) MongoGetForeignPaths
and MongoGetForeignPlan functions (based on file_fdw).