Fix build under Postgres 9.3
With the change to the `.SECONDARY` rule in [this commit][1], vanilla
Postgres extension builds do not keep intermediate files.

Rather than directly defining `all` and `clean` targets, we should be
following the [Postgres guidelines for making extensions][2] if at all
possible. This is as simple as defining special variables before in-
cluding the PGXS Makefile and making sure all relevant dependencies can
be built if needed.

[1]: postgres/postgres@1eb1dde
[2]: http://www.postgresql.org/docs/current/static/extend-pgxs.html