Error (#32) full document retrieval.
This feature allows the user to retrieve the
full mongo document in a single JSON column.
The user would need to add a __doc json column
to the foreign table and mongo_fdw will fetch
the entire mongo document in this column.This
feature only supported in MongoC Meta driver.