Check that the existing or new connection is reachable/active or not.
mongoc_client_new() function does not return NULL when it fails to
connect with the server.  So we need to ping the server to make sure
that connection is reachable with the parameters given.

Along the way, move the connection establishment logic from
MongoExecForeignInsert/Update/Delete functions to the
MongoBeginForeignModify() function so that we won't try to ping the
server for every insert/update/delete operations.

FDW-127, Vaibhav Dalvi, reviewed by Suraj Kharage.