Change the server port option's type from int32 to uint16.
Meta driver APIs expect port value in unsigned short type which
resulted in a compilation warning on some gcc version as we were
passing an int to it.  So change that to fix the compilation warnings.
The legacy driver, however, accepts an int, so we cast that value to
int while passing to those APIs.  Now as we have changed the type to
uint16, added a strict check making sure that the value for the port
provided is within the ushort range.

FDW-199, Vaibhav Dalvi, reviewed and test-case added by me.