Fix ReScanForeignScan API to make the parameterized query work correctly.
Sub-select or correlated queries that use a parameterized plan
uses ReScanForeignScan API.  However, this API is not correctly
rescanning the data.  In the case of parameters, we need to
recreate the Mongo cursor that fetches the data again.  The
patch does the same.

In passing, refactor some code in this area so that Mongo cursor
is created from MongoIterateForeignScan() that is called after
Begin and ReScan APIs.  However, create a Mongo cursor only if
it was not created already.

FDW-103, Vaibhav Dalvi, reviewed by Suraj Kharage.