Fix issue where casting target list produces 'NULL'.
To check which columns are needed to be output, we create a column
list from the target list.  However, we are only adding the plain
Var nodes appearing in the target list.  Thus, if a target list has
an explicit casting, then it won't appear as a plain Var node and
thus not added in the column list, resulting in returning NULL for
them.  Fix that by correctly pulling out the Vars from the expression
appearing in the target list by using pull_var_clause().  This enables
us to give a correct result even if we have expressions other than an
explicit cast like a function call or operators.

Reported on GitHub through issues #133 by Maksim Volkau (dadhi).

FDW-197, Vaibhav Dalvi, reviewed by Suraj Kharage