Rfactor WHERE pushdown code.
Earlier, we are finding the pushable OpExpr by calling the function
ApplicableOpExpressionList().  Change that to the postgres_fdw
way of having a mongo_is_foreign_expr() function and a
foreign_expr_walker().  This is cleaner and in-line with PostgreSQL
that can be extended easily.  This refactoring allowed is to push
down the clauses having a Var node wrapped in the RelabelType.

Also, for no reason but just for simplicity, we were re-evaluating
restriction clauses that are pushed down to MongoDB again on the
local side.  Remove that altogether which adds to some performance
improvement.

FDW-194, Vaibhav Dalvi, reviewed by Suraj Kharage, further
improvements by Jeevan Chalke.