Fall back to cmake if cmake3 is not present.
The autogen.sh is using cmake3 to compile the jsonc library and
mongoc driver.  However, it is observed that some of the platforms
do not have that, and thus script fails.  So, if cmake3 is not present
on the system, then fall back to cmake and use that.  Note that if the
CMake version is below the required one (i.e. 3.1), then the script
will fail.

Reported on GitHub through issues #149 by
Andreas Scherbaum (andreasscherbaum).

FDW-379, Vaibhav Dalvi, reviewed by Jeevan Ladhe,
tested by Rajkumar Raghuwanshi.