Make regression consistent on rhel8 by changing the host details.
For invalid host and port testing, we were using the IP address
"127.0.0.5" (with default Mongo port).  However, on the rhel8
platform, we observed that the same IP is used as loopback IP for
the localhost and thus was not testing the stuff we wanted to test
and resulted in a regression.  So to avoid failure on such platforms,
change the address to "127.0.0.10" and port to 9999 instead.

FDW-423, Rajkumar Raghuwanshi, reviewed and further fixed by
Vaibhav Dalvi, tested by Varsha Mehtre.