Add new server option use_remote_estimate.
We were estimating the rows by calling ForeignTableDocumentCount()
which is costly, and there is no way to avoid that.  Thus add a new
server-level option use_remote_estimate so that we will estimate the
actual rows only when it is set to true.  Default is false.

FDW-464, Vaibhav Dalvi, reviewed by Suraj Kharage, tested by
Kashif Zeeshan.