Fix server crash while updating/deleting with null _id value.
For update/delete operations, we rely on a unique value, and for that,
we use the "_id" column, which is supposed to be non-null.  However,
the user can insert a NULL value to this column on the MongoDB side.
And as we don't have handling for the same, we ended up with a server
crash along the way.  Fix it by correctly appending the NULL value to
this column in the update/delete query.

Along the way, tighten the checks around the "_id" column.  We enforce
that the first column of the table must be named "_id" and of type
"NAME".  Also, on the MongoDB side, it should be of the "ObjectId"
type.

Moreover, we are not inserting any value for the "_id" column by
ignoring the passed-in value.  Since the MongoDB server inserts the
unique value for the missing "_id" column, we are ignoring it.  Emit
the debug1 log message for the same.

FDW-481, Vaibhav Dalvi, reviewed by Suraj Kharage and me,
tested by Ajay Pal.