Add enable_aggregate_pushdown option to control aggregate push-down.
Considering the complexity of generating the aggregation pipeline,
add an option to control the push-down to avoid any issues.

This option can be set at the server level or the table level. If any
of the tables involved in the join has this option set to false, then
the join will not be pushed down.  The table-level value of the option
takes precedence over the server-level option value.  Default is true.

FDW-137, Vaibhav Dalvi, reviewed by Suraj Kharage and Jeevan Chalke,
tested by Kashif Zeeshan.