Add deprecation notice for the legacy driver in README and autogen.sh.
The legacy driver is too old, and MongoDB also stopped supporting it.
Given that all the new feature needs a meta driver, having a legacy
driver leads to complexity.  Add a warning for the same in autogen.sh
if someone tries building it, and also, put a note in the README.

FDW-471, Vaibhav Dalvi, reviewed by Suraj Kharage.