Fix server crash caused due to missed Param node handling.
While preparing a query filter in mongo_query_document(), we have
missed considering the Param node in a comparison expression.  And
thus, was not preparing the query correctly, leading to the server
crash.  We do have this handling for the equality operator, though.
Fix the same.

Reported on GitHub through issue #160 by pludov.

FDW-529, Vaibhav Dalvi, reviewed and further revised by Jeevan Chalke.