Save the status of the enable_aggregate_pushdown option for later checking.
In the case of nested join queries, the existing logic may not
correctly propagate the status of the enable_aggregate_pushdown
option to the stage where we are accessing the aggregation path.
Fix that by saving the status of enable_aggregate_pushdown in
MongoFdwRelationInfo, which is updated for every output rel at
various stages. It can then be accessed at the time of grouping
operation.

FDW-560, Vaibhav Dalvi, reviewed by Sravan Velagandula, tested by
Kashif Zeeshan.